+++ packages/components/src/ProForm/widgets/Upload/types.ts
@@ -1,0 +1,43 @@
+import React from 'react';
+
+export interface ImageType {
+  dataURL?: string;
+  file?: File;
+  [key: string]: any;
+}
+
+export type ImageListType = Array<ImageType>;
+
+export interface ImageUploadingPropsType {
+  value: ImageListType;
+  onChange: (value: ImageListType, addUpdatedIndex?: Array<number>) => void;
+  children?: (props: ExportInterface) => React.ReactNode;
+  multiple?: boolean;
+  maxNumber?: number;
+  accept?: Array<string>;
+  maxFileSize?: number;
+  resolutionWidth?: number;
+  resolutionHeight?: number;
+  resolutionType?: ResolutionType;
+  onError?: (errors: ErrorsType, files?: ImageListType) => void;
+  dataURLKey?: string;
+  inputProps?: React.HTMLProps<HTMLInputElement>;
+}
+
+export interface ExportInterface {
+  imageList: ImageListType;
+  onImageUpload: () => void;
+  onImageRemoveAll: () => void;
+  errors: ErrorsType;
+  onImageUpdate: (index: number) => void;
+  onImageRemove: (index: number) => void;
+}
+
+export type ErrorsType = {
+  maxFileSize?: boolean;
+  maxNumber?: boolean;
+  accept?: boolean;
+  resolution?: boolean;
+} | null;
+
+export type ResolutionType = 'absolute' | 'less' | 'more' | 'ratio';